+++ batch/src/main/java/com/yachaerang/batch/scheduler/MonthlyPriceScheduler.java
@@ -1,7 +1,7 @@
 package com.yachaerang.batch.scheduler;
 
 import com.yachaerang.batch.domain.dto.JobStartResponseDto;
-import com.yachaerang.batch.service.BatchJobService;
+import com.yachaerang.batch.service.job.BatchJobService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.batch.core.JobExecutionException;
